chore: move isXMLFile function from scan to config
Plugin is configured through ComplyTime where a validation was
implemented. So, this validation was also moved to config in order to
report potential issues at an early stage.
